code/requirements.txt
pyserial
python-telegram-bot
emoji
opencv-python
matplotlib
pandas
